refactor: 각 컴포넌트에서 UI 그리는 함수와 이벤트핸들러를 분리
