+++ README.md
@@ -19,6 +19,9 @@
   - ColumnSection
 - 공통되지 않더라도 관심사별로 분리하기
   - Header
+- 컴포넌트 역할별로 분리하고 중복되는 요소 공용화
+- 데이터를 여러 컴포넌트에서 공유하기 위해 객체를 넘겨 값을 변환
+- 각 컴포넌트에서 UI 그리는 함수와 이벤트핸들러를 분리
 
 ## 코멘트 모음
