+++ components/ColumnBody/ColumnBody.js
@@ -1,0 +1,7 @@
+import createColumnBody from "./ui/createColumnBody.js";
+
+const ColumnBody = ({ items }) => {
+  return createColumnBody({ items });
+};
+
+export default ColumnBody;
