+++ components/ColumnHeader/ColumnHeader.js
@@ -1,0 +1,37 @@
+import ColumnInputItem from "../ColumnInputItem/ColumnInputItem.js";
+import createColumnHeader from "./ui/createColumnHeader.js";
+
+const ColumnHeader = ({ id, title, items, store }) => {
+  const handleCancel = (e) => {
+    const target = e.target;
+    const $columnItem = target
+      .closest(".column__container")
+      .querySelector(".column__body");
+
+    $columnItem.firstChild.remove();
+    store.isTodoAdding = false;
+  };
+
+  const handleClickAdd = (e) => {
+    if (store.isTodoAdding) {
+      handleCancel(e);
+    } else {
+      addColumnItemInput({ store });
+    }
+  };
+
+  const addColumnItemInput = ({ store }) => {
+    const $columnBody = document.querySelector(`#${id} .column__body`);
+    const $columnItemInput = ColumnInputItem({ store, handleCancel });
+    $columnBody.prepend($columnItemInput);
+    store.isTodoAdding = true;
+  };
+
+  return createColumnHeader({
+    title,
+    count: items.length,
+    handleClickAdd,
+  });
+};
+
+export default ColumnHeader;
