+++ components/ColumnSection/ColumnSection.js
@@ -1,0 +1,9 @@
+import createColumnSection from "./ui/createColumnSection.js";
+
+const ColumnSection = ({ id, title, items }) => {
+  const store = { isTodoAdding: false };
+
+  return createColumnSection({ id, title, items, store });
+};
+
+export default ColumnSection;
