+++ components/Header/Header.js
@@ -1,0 +1,7 @@
+import createHeader from "./ui/createHeader.js";
+
+const Header = () => {
+  return createHeader();
+};
+
+export default Header;
